downgrade pytorch lightning to 1.9.5
